New Line code (W.I.P)
Here is some of the line code right now. Test.html has the line code that I was trying to get to work using the image. The image requires different kinds of parameters to work it seems. Also I noticed in the test.html it says it can support undo and redo feature going to see if that will fix the current undo and redo problem.